Included Bundling Path names as string for YearActivity, SemesterActivity, and CriteriaActivity.YearActivity passes the yearPath in a bundle as a string to SemesterActivity
SemesterActivity receives the yearPath from the bundle from
YearActivity, It also sends a bundle to CourseActivity which includes
yearPath and semesterPath.

CourseActivity receives a bundle from SemesterActivity which includes
yearPath and SemesterPath. It also sends a bundle to CriteriaActivity
which includes yearPath, semesterPath, and coursePath.

CriteriaActivity receives the yearPath, semesterPath, coursePath from
the bundle passed to it by CourseActivity.

I also included new methods in each class to bundle and pass back the
appropriate information for each transition when the TextView is click
in order to return to the appropriate Activity.

No further bundles have been implemented at this point.